// src/types/types.ts
export type OpenAIModel = 'gpt-3.5-turbo' | 'gpt-4-1106-preview'| 'gpt-4o';
// ...
export interface CaptionBody {
  topic: string;
  toneOfVoice:
    | ''
    | 'Formal'
    | 'Informal'
    | 'Humorous'
    | 'Serious'
    | 'Optimistic'
    | 'Motivating'
    | 'Respectful'
    | 'Assertive'
    | 'Conversational';
  description: string;
  model: OpenAIModel;
  apiKey?: string | undefined | null;
}